--- src/components/Login/Login.css
@@ -1,126 +1,0 @@
-.Login{
-    display: flex;
-    flex-direction: column; /* 요소들을 세로로 배치 */
-    justify-content: center; /* 가로 방향 중앙 정렬 */
-    align-items: center; /* 세로 방향 중앙 정렬 */
-    
-    height: 100vh; 
-
-    .title{
-        margin-bottom: 15px;
-        text-align: left;
-    }
-
-        .Login_Title{
-            text-decoration: none;
-            font-size: 30px;
-            color: rgb(43, 43, 43);
-            font-family: 'Tenada';
-        }
-
-        .Login_Text{
-            font-size: 17px;
-            font-family: 'ChosunGu';
-            font-weight: 500;
-        }
-
-
-    .Login_Box{
-        display: flex;
-        flex-direction: column; /* 요소들을 세로로 배치 */
-        justify-content: center; /* 가로 방향 중앙 정렬 */
-        align-items: center; /* 세로 방향 중앙 정렬 */
-
-        margin-bottom: 20px;
-    }
-
-        .Login_ID, .Login_PW{
-            width: 15em;
-            border-top: none;
-            border-left: none;
-            border-right: none;
-            border-bottom: 1px solid rgb(43, 43, 43);
-
-            height: 30px;
-
-            margin-top: 1em;
-
-            color: rgb(51, 51, 51);
-            font-size: 16px;
-
-            
-        }
-
-        .Login_btn{
-            width: 18.5em;
-            height: 33px;
-            margin-top: 1.3em;
-
-            color: aliceblue;
-            background-color: rgb(43, 43, 43);
-
-            border-radius: 10px;
-            border-style: none;
-
-            font-family: 'ChosunGu';
-
-            box-shadow: 0px 4px 1px black;
-        }
-        .Login_btn:hover{
-            background-color: rgb(255, 76, 100);
-        }
-        .Login_btn:active{
-            border: 1px solid rgb(43, 43, 43);
-            background-color: aliceblue;
-
-            color: rgb(43, 43, 43);
-            width: 18em;
-        }
-
-        .Login_Other{
-            .Login_LostPw{
-                margin-right: 1em;
-                font-size: 16px;
-                font-weight: 500;
-                font-family: 'ChosunGu';
-                text-decoration: none;
-
-                color: rgb(43, 43, 43);
-            }
-
-            .Login_SignUp{
-                margin-left: 1em;
-                font-size: 16px;
-                font-weight: 500;
-                font-family: 'ChosunGu';
-                text-decoration: none;
-
-                color: rgb(43, 43, 43);
-            }
-        }
-    
-
-
-
-}
-
-
-
-
-
-
-
-@font-face {
-    font-family: 'Tenada';
-    src: url('https://fastly.jsdelivr.net/gh/projectnoonnu/noonfonts_2210-2@1.0/Tenada.woff2') format('woff2');
-    font-weight: normal;
-    font-style: normal;
-}
-
-
-@font-face {
-    font-family: 'ChosunGu';
-    src: url('https://fastly.jsdelivr.net/gh/projectnoonnu/noonfonts_20-04@1.0/ChosunGu.woff') format('woff');
-    font-weight: normal;
-    font-style: normal;
-}
